requirements.txt
# requirements.txt
flask
google-genai
pytest
pytest-mock